Refactoring + update deps
