+++ demo/package.json
@@ -7,11 +7,11 @@
   },
   "dependencies": {
     "autoprefixer": "^10.4.18",
-    "next": "^14.1.2",
+    "next": "^15.5.2",
     "postcss": "^8.4.35",
     "tailwindcss": "^3.4.1"
   },
   "devDependencies": {
-    "@types/node": "20.11.24"
+    "@types/node": "22"
   }
 }
