+++ demo/src/app/page.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { NextPage } from "next";
 import dynamic from "next/dynamic";
 const AdminApp = dynamic(() => import("@/components/AdminApp"), { ssr: false });
