+++ demo/src/components/AdminApp.tsx
@@ -2,11 +2,15 @@
 import { dataProvider } from "@/app/dataProvider";
 import { Admin, Resource } from "react-admin";
 import { PagesEdit, PagesList } from "./ResourcePage";
+import { DndProvider } from "react-dnd";
+import { HTML5Backend } from "react-dnd-html5-backend";
 
 const AdminApp = () => (
-  <Admin dataProvider={dataProvider}>
-    <Resource name="pages" list={PagesList} edit={PagesEdit} />
-  </Admin>
+  <DndProvider backend={HTML5Backend}>
+    <Admin dataProvider={dataProvider}>
+      <Resource name="pages" list={PagesList} edit={PagesEdit} />
+    </Admin>
+  </DndProvider>
 );
 
 export default AdminApp;
