+++ package.json
@@ -26,37 +26,42 @@
     "demo": "cd demo && yarn dev"
   },
   "devDependencies": {
-    "@mui/icons-material": "5",
-    "@mui/material": "5",
-    "@types/react": "^18.2.63",
+    "@mui/icons-material": "7",
+    "@mui/material": "7",
+    "@mui/styles": "^6.4.8",
+    "@types/react": "^19",
     "@typescript-eslint/eslint-plugin": "^7.1.1",
     "@typescript-eslint/parser": "^7.1.1",
+    "dnd-core": "^16",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-react": "^7.34.0",
-    "next": "^14.1.2",
+    "next": "^15.5.2",
     "prettier": "^3.2.5",
-    "ra-data-fakerest": "^4.16.12",
-    "ra-data-json-server": "^4.16.12",
-    "react": "^18.2.0",
-    "react-admin": "^4.16.12",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.51.0",
-    "tsup": "^8.0.2",
-    "typescript": "^4.5.5"
+    "ra-data-fakerest": "^5.11.0",
+    "ra-data-json-server": "^5.11.0",
+    "react": "^19",
+    "react-admin": "^5.11.0",
+    "react-dnd": "^16",
+    "react-dnd-html5-backend": "^16",
+    "react-dom": "^19",
+    "react-hook-form": "^7.62.0",
+    "tsup": "^8.5.0",
+    "typescript": "^5.9.2"
   },
   "dependencies": {
     "clsx": "^2",
-    "dnd-core": "^16",
     "nanoid": "^5",
-    "react-dnd": "^16",
-    "react-dnd-html5-backend": "^16",
     "react-resizable-panels": "^2"
   },
   "peerDependencies": {
-    "@mui/icons-material": "^5",
-    "react": "^18",
-    "react-dom": "^18"
+    "@mui/icons-material": "^7",
+    "@mui/material": "^7",
+    "dnd-core": "^16",
+    "react": "^18 || ^19",
+    "react-dnd": "^16",
+    "react-dnd-html5-backend": "^16",
+    "react-dom": "^18 || ^19"
   },
   "resolutions": {
     "strip-ansi": "6.0.1"
