+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./BlocksInput";
+export * from "./BlocksInputWhitoutDndContext";
 export * from "./types";
